move types to types.ts + add coinSelector file
